+++ internal/sbi/api_httpcallback.go
@@ -1,6 +1,15 @@
 package sbi
 
-import "strings"
+import (
+	"net/http"
+	"strings"
+
+	"github.com/gin-gonic/gin"
+
+	"github.com/free5gc/openapi"
+	"github.com/free5gc/openapi/models"
+	"github.com/free5gc/udm/internal/logger"
+)
 
 func (s *Server) getHttpCallBackRoutes() []Route {
 	return []Route{
@@ -8,14 +17,60 @@ func (s *Server) getHttpCallBackRoutes() []Route {
 			"Index",
 			"GET",
 			"/",
-			s.Processor().HandleIndex,
+			s.HandleIndex,
 		},
 
 		{
 			"DataChangeNotificationToNF",
 			strings.ToUpper("Post"),
 			"/sdm-subscriptions",
-			s.Processor().HandleDataChangeNotificationToNF,
+			s.HandleDataChangeNotificationToNF,
 		},
+	}
+}
+
+func (s *Server) HandleDataChangeNotificationToNF(c *gin.Context) {
+	var dataChangeNotify models.DataChangeNotify
+	// step 1: retrieve http request body
+	requestBody, err := c.GetRawData()
+	if err != nil {
+		problemDetail := models.ProblemDetails{
+			Title:  "System failure",
+			Status: http.StatusInternalServerError,
+			Detail: err.Error(),
+			Cause:  "SYSTEM_FAILURE",
+		}
+		logger.CallbackLog.Errorf("Get Request Body error: %+v", err)
+		c.JSON(http.StatusInternalServerError, problemDetail)
+		return
+	}
+
+	// step 2: convert requestBody to openapi models
+	err = openapi.Deserialize(&dataChangeNotify, requestBody, "application/json")
+	if err != nil {
+		problemDetail := "[Request Body] " + err.Error()
+		rsp := models.ProblemDetails{
+			Title:  "Malformed request syntax",
+			Status: http.StatusBadRequest,
+			Detail: problemDetail,
+		}
+		logger.CallbackLog.Errorln(problemDetail)
+		c.JSON(http.StatusBadRequest, rsp)
+		return
+	}
+
+	supi := c.Params.ByName("supi")
+
+	logger.CallbackLog.Infof("Handle DataChangeNotificationToNF")
+
+	problemDetails := s.Processor().DataChangeNotificationProcedure(dataChangeNotify.NotifyItems, supi)
+
+	// step 4: process the return value from step 3
+	if problemDetails != nil {
+		c.JSON(int(problemDetails.Status), problemDetails)
+		return
+	} else {
+		c.Status(http.StatusNoContent)
+		return
 	}
 }
