+++ internal/sbi/api_parameterprovision.go
@@ -1,6 +1,15 @@
 package sbi
 
-import "strings"
+import (
+	"net/http"
+	"strings"
+
+	"github.com/gin-gonic/gin"
+
+	"github.com/free5gc/openapi"
+	"github.com/free5gc/openapi/models"
+	"github.com/free5gc/udm/internal/logger"
+)
 
 func (s *Server) getParameterProvisionRoutes() []Route {
 	return []Route{
@@ -8,14 +17,62 @@ func (s *Server) getParameterProvisionRoutes() []Route {
 			"Index",
 			"GET",
 			"/",
-			s.Processor().HandleIndex,
+			s.HandleIndex,
 		},
 
 		{
 			"Update",
 			strings.ToUpper("Patch"),
 			"/:gpsi/pp-data",
-			s.Processor().HandleUpdate,
+			s.HandleUpdate,
 		},
+	}
+}
+
+func (s *Server) HandleUpdate(c *gin.Context) {
+	var ppDataReq models.PpData
+
+	// step 1: retrieve http request body
+	requestBody, err := c.GetRawData()
+	if err != nil {
+		problemDetail := models.ProblemDetails{
+			Title:  "System failure",
+			Status: http.StatusInternalServerError,
+			Detail: err.Error(),
+			Cause:  "SYSTEM_FAILURE",
+		}
+		logger.PpLog.Errorf("Get Request Body error: %+v", err)
+		c.JSON(http.StatusInternalServerError, problemDetail)
+		return
+	}
+
+	// step 2: convert requestBody to openapi models
+	err = openapi.Deserialize(&ppDataReq, requestBody, "application/json")
+	if err != nil {
+		problemDetail := "[Request Body] " + err.Error()
+		rsp := models.ProblemDetails{
+			Title:  "Malformed request syntax",
+			Status: http.StatusBadRequest,
+			Detail: problemDetail,
+		}
+		logger.PpLog.Errorln(problemDetail)
+		c.JSON(http.StatusBadRequest, rsp)
+		return
+	}
+
+	gpsi := c.Params.ByName("gpsi")
+
+	logger.PpLog.Infoln("Handle UpdateRequest")
+
+	// step 3: handle the message
+	problemDetails := s.Processor().UpdateProcedure(ppDataReq, gpsi)
+
+	// step 4: process the return value from step 3
+	if problemDetails != nil {
+		c.JSON(int(problemDetails.Status), problemDetails)
+		return
+	} else {
+		c.Status(http.StatusNoContent)
+		return
 	}
 }
