+++ internal/sbi/api_ueauthentication.go
@@ -1,6 +1,15 @@
 package sbi
 
-import "strings"
+import (
+	"net/http"
+	"strings"
+
+	"github.com/gin-gonic/gin"
+
+	"github.com/free5gc/openapi"
+	"github.com/free5gc/openapi/models"
+	"github.com/free5gc/udm/internal/logger"
+)
 
 func (s *Server) getUEAuthenticationRoutes() []Route {
 	return []Route{
@@ -8,14 +17,129 @@ func (s *Server) getUEAuthenticationRoutes() []Route {
 			"Index",
 			"GET",
 			"/",
-			s.Processor().HandleIndex,
+			s.HandleIndex,
 		},
 
 		{
 			"ConfirmAuth",
 			strings.ToUpper("Post"),
 			"/:supi/auth-events",
-			s.Processor().HandleConfirmAuth,
+			s.HandleConfirmAuth,
 		},
 	}
+}
+
+// ConfirmAuth - Create a new confirmation event
+func (s *Server) HandleConfirmAuth(c *gin.Context) {
+	var authEvent models.AuthEvent
+	// step 1: retrieve http request body
+	requestBody, err := c.GetRawData()
+	if err != nil {
+		problemDetail := models.ProblemDetails{
+			Title:  "System failure",
+			Status: http.StatusInternalServerError,
+			Detail: err.Error(),
+			Cause:  "SYSTEM_FAILURE",
+		}
+		logger.UeauLog.Errorf("Get Request Body error: %+v", err)
+		c.JSON(http.StatusInternalServerError, problemDetail)
+		return
+	}
+
+	// step 2: convert requestBody to openapi models
+	err = openapi.Deserialize(&authEvent, requestBody, "application/json")
+	if err != nil {
+		problemDetail := "[Request Body] " + err.Error()
+		rsp := models.ProblemDetails{
+			Title:  "Malformed request syntax",
+			Status: http.StatusBadRequest,
+			Detail: problemDetail,
+		}
+		logger.UeauLog.Errorln(problemDetail)
+		c.JSON(http.StatusBadRequest, rsp)
+		return
+	}
+
+	supi := c.Params.ByName("supi")
+
+	logger.UeauLog.Infoln("Handle ConfirmAuthDataRequest")
+
+	problemDetails := s.Processor().ConfirmAuthDataProcedure(authEvent, supi)
+
+	if problemDetails != nil {
+		c.JSON(int(problemDetails.Status), problemDetails)
+		return
+	} else {
+		c.Status(http.StatusCreated)
+		return
+	}
+}
+
+// GenerateAuthData - Generate authentication data for the UE
+func (s *Server) HandleGenerateAuthData(c *gin.Context) {
+	var authInfoReq models.AuthenticationInfoRequest
+
+	// step 1: retrieve http request body
+	requestBody, err := c.GetRawData()
+	if err != nil {
+		problemDetail := models.ProblemDetails{
+			Title:  "System failure",
+			Status: http.StatusInternalServerError,
+			Detail: err.Error(),
+			Cause:  "SYSTEM_FAILURE",
+		}
+		logger.UeauLog.Errorf("Get Request Body error: %+v", err)
+		c.JSON(http.StatusInternalServerError, problemDetail)
+		return
+	}
+
+	// step 2: convert requestBody to openapi models
+	err = openapi.Deserialize(&authInfoReq, requestBody, "application/json")
+	if err != nil {
+		problemDetail := "[Request Body] " + err.Error()
+		rsp := models.ProblemDetails{
+			Title:  "Malformed request syntax",
+			Status: http.StatusBadRequest,
+			Detail: problemDetail,
+		}
+		logger.UeauLog.Errorln(problemDetail)
+		c.JSON(http.StatusBadRequest, rsp)
+		return
+	}
+
+	// step 1: log
+	logger.UeauLog.Infoln("Handle GenerateAuthDataRequest")
+
+	// step 2: retrieve request
+	supiOrSuci := c.Param("supiOrSuci")
+
+	// step 3: handle the message
+	response, problemDetails := s.Processor().GenerateAuthDataProcedure(authInfoReq, supiOrSuci)
+
+	// step 4: process the return value from step 3
+	if response != nil {
+		// status code is based on SPEC, and option headers
+		c.JSON(http.StatusOK, response)
+		return
+	} else if problemDetails != nil {
+		c.JSON(int(problemDetails.Status), problemDetails)
+		return
+	} else {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  "UNSPECIFIED",
+		}
+		c.JSON(http.StatusForbidden, problemDetails)
+		return
+	}
+}
+
+func (s *Server) GenAuthDataHandlerFunc(c *gin.Context) {
+	c.Params = append(c.Params, gin.Param{Key: "supiOrSuci", Value: c.Param("supi")})
+	if strings.ToUpper("Post") == c.Request.Method {
+		s.HandleGenerateAuthData(c)
+		return
+	}
+
+	c.String(http.StatusNotFound, "404 page not found")
 }
