--- internal/sbi/processor/amf_non3_gpp_access_registration_info_retrieval.go
@@ -1,45 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/antihax/optional"
-	"github.com/gin-gonic/gin"
-
-	"github.com/free5gc/openapi/Nudr_DataRepository"
-	"github.com/free5gc/openapi/models"
-	"github.com/free5gc/udm/internal/logger"
-)
-
-// GetAmfNon3gppAccess - retrieve the AMF registration for non-3GPP access information
-func (p *Processor) HandleGetAmfNon3gppAccess(c *gin.Context) {
-	// step 1: log
-	logger.UecmLog.Infoln("Handle GetAmfNon3gppAccessRequest")
-
-	// step 2: retrieve request
-	ueId := c.Param("ueId")
-	supportedFeatures := c.Query("supported-features")
-
-	var queryAmfContextNon3gppParamOpts Nudr_DataRepository.QueryAmfContextNon3gppParamOpts
-	queryAmfContextNon3gppParamOpts.SupportedFeatures = optional.NewString(supportedFeatures)
-	// step 3: handle the message
-
-	response, problemDetails := p.consumer.GetAmfNon3gppAccessProcedure(queryAmfContextNon3gppParamOpts, ueId)
-
-	// step 4: process the return value from step 3
-	if response != nil {
-		// status code is based on SPEC, and option headers
-		c.JSON(http.StatusOK, response)
-		return
-	} else if problemDetails != nil {
-		c.JSON(int(problemDetails.Status), problemDetails)
-		return
-	} else {
-		problemDetails = &models.ProblemDetails{
-			Status: http.StatusForbidden,
-			Cause:  "UNSPECIFIED",
-		}
-		c.JSON(http.StatusForbidden, problemDetails)
-		return
-	}
-}
