--- internal/sbi/processor/amf_registration_for3_gpp_access.go
@@ -1,73 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-
-	"github.com/free5gc/openapi"
-	"github.com/free5gc/openapi/models"
-	"github.com/free5gc/udm/internal/logger"
-)
-
-// RegistrationAmf3gppAccess - register as AMF for 3GPP access
-func (p *Processor) HandleRegistrationAmf3gppAccess(c *gin.Context) {
-	var amf3GppAccessRegistration models.Amf3GppAccessRegistration
-	// step 1: retrieve http request body
-	requestBody, err := c.GetRawData()
-	if err != nil {
-		problemDetail := models.ProblemDetails{
-			Title:  "System failure",
-			Status: http.StatusInternalServerError,
-			Detail: err.Error(),
-			Cause:  "SYSTEM_FAILURE",
-		}
-		logger.UecmLog.Errorf("Get Request Body error: %+v", err)
-		c.JSON(http.StatusInternalServerError, problemDetail)
-		return
-	}
-
-	// step 2: convert requestBody to openapi models
-	err = openapi.Deserialize(&amf3GppAccessRegistration, requestBody, "application/json")
-	if err != nil {
-		problemDetail := "[Request Body] " + err.Error()
-		rsp := models.ProblemDetails{
-			Title:  "Malformed request syntax",
-			Status: http.StatusBadRequest,
-			Detail: problemDetail,
-		}
-		logger.UecmLog.Errorln(problemDetail)
-		c.JSON(http.StatusBadRequest, rsp)
-		return
-	}
-
-	// step 1: log
-	logger.UecmLog.Infof("Handle RegistrationAmf3gppAccess")
-
-	// step 2: retrieve request
-	ueID := c.Param("ueId")
-	logger.UecmLog.Info("UEID: ", ueID)
-
-	// step 3: handle the message
-	header, response, problemDetails := p.consumer.RegistrationAmf3gppAccessProcedure(amf3GppAccessRegistration, ueID)
-
-	// step 4: process the return value from step 3
-	if response != nil {
-		if header != nil {
-			// status code is based on SPEC, and option headers
-			for key, val := range header { // header response is optional
-				c.Header(key, val[0])
-			}
-			c.JSON(http.StatusCreated, response)
-			return
-		}
-		c.JSON(http.StatusOK, response)
-		return
-	} else if problemDetails != nil {
-		c.JSON(int(problemDetails.Status), problemDetails)
-		return
-	} else {
-		c.Status(http.StatusNoContent)
-		return
-	}
-}
