+++ internal/sbi/processor/generate_auth_data.go
@@ -1,70 +1,553 @@
 package processor
 
 import (
+	cryptoRand "crypto/rand"
+	"encoding/hex"
+	"fmt"
+	"math/big"
+	"math/rand"
 	"net/http"
+	"reflect"
+	"strconv"
+	"strings"
+	"time"
 
-	"github.com/gin-gonic/gin"
+	"github.com/antihax/optional"
 
 	"github.com/free5gc/openapi"
+	"github.com/free5gc/openapi/Nudr_DataRepository"
 	"github.com/free5gc/openapi/models"
+	udm_context "github.com/free5gc/udm/internal/context"
 	"github.com/free5gc/udm/internal/logger"
+	"github.com/free5gc/udm/pkg/suci"
+	"github.com/free5gc/util/milenage"
+	"github.com/free5gc/util/ueauth"
 )
 
-// GenerateAuthData - Generate authentication data for the UE
-func (p *Processor) HandleGenerateAuthData(c *gin.Context) {
-	var authInfoReq models.AuthenticationInfoRequest
+const (
+	SqnMAx    int64 = 0xFFFFFFFFFFFF
+	ind       int64 = 32
+	keyStrLen int   = 32
+	opStrLen  int   = 32
+	opcStrLen int   = 32
+)
 
-	// step 1: retrieve http request body
-	requestBody, err := c.GetRawData()
+const (
+	authenticationRejected string = "AUTHENTICATION_REJECTED"
+	resyncAMF              string = "0000"
+)
+
+func (p *Processor) aucSQN(opc, k, auts, rand []byte) ([]byte, []byte) {
+	AK, SQNms := make([]byte, 6), make([]byte, 6)
+	macS := make([]byte, 8)
+	ConcSQNms := auts[:6]
+	AMF, err := hex.DecodeString(resyncAMF)
 	if err != nil {
-		problemDetail := models.ProblemDetails{
-			Title:  "System failure",
-			Status: http.StatusInternalServerError,
+		return nil, nil
+	}
+
+	logger.UeauLog.Tracef("aucSQN: ConcSQNms=[%x]", ConcSQNms)
+
+	err = milenage.F2345(opc, k, rand, nil, nil, nil, nil, AK)
+	if err != nil {
+		logger.UeauLog.Errorln("aucSQN milenage F2345 err:", err)
+	}
+
+	for i := 0; i < 6; i++ {
+		SQNms[i] = AK[i] ^ ConcSQNms[i]
+	}
+
+	logger.UeauLog.Tracef("aucSQN: opc=[%x], k=[%x], rand=[%x], AMF=[%x], SQNms=[%x]\n", opc, k, rand, AMF, SQNms)
+	// The AMF used to calculate MAC-S assumes a dummy value of all zeros
+	err = milenage.F1(opc, k, rand, SQNms, AMF, nil, macS)
+	if err != nil {
+		logger.UeauLog.Errorln("aucSQN milenage F1 err:", err)
+	}
+	logger.UeauLog.Tracef("aucSQN: macS=[%x]\n", macS)
+	return SQNms, macS
+}
+
+func (p *Processor) strictHex(ss string, n int) string {
+	l := len(ss)
+	if l < n {
+		return strings.Repeat("0", n-l) + ss
+	} else {
+		return ss[l-n : l]
+	}
+}
+
+func (p *Processor) ConfirmAuthDataProcedure(
+	authEvent models.AuthEvent,
+	supi string,
+) (problemDetails *models.ProblemDetails) {
+	ctx, pd, err := udm_context.GetSelf().GetTokenCtx(models.ServiceName_NUDR_DR, models.NfType_UDR)
+	if err != nil {
+		return pd
+	}
+	var createAuthParam Nudr_DataRepository.CreateAuthenticationStatusParamOpts
+	optInterface := optional.NewInterface(authEvent)
+	createAuthParam.AuthEvent = optInterface
+
+	client, err := p.consumer.CreateUDMClientToUDR(supi)
+	if err != nil {
+		return openapi.ProblemDetailsSystemFailure(err.Error())
+	}
+
+	resp, err := client.AuthenticationStatusDocumentApi.CreateAuthenticationStatus(
+		ctx, supi, &createAuthParam)
+	if err != nil {
+		problemDetails = &models.ProblemDetails{
+			Status: int32(resp.StatusCode),
+			Cause:  err.(openapi.GenericOpenAPIError).Model().(models.ProblemDetails).Cause,
 			Detail: err.Error(),
-			Cause:  "SYSTEM_FAILURE",
 		}
-		logger.UeauLog.Errorf("Get Request Body error: %+v", err)
-		c.JSON(http.StatusInternalServerError, problemDetail)
-		return
-	}
 
-	// step 2: convert requestBody to openapi models
-	err = openapi.Deserialize(&authInfoReq, requestBody, "application/json")
+		logger.UeauLog.Errorln("ConfirmAuth err:", err.Error())
+		return problemDetails
+	}
+	defer func() {
+		if rspCloseErr := resp.Body.Close(); rspCloseErr != nil {
+			logger.UeauLog.Errorf("CreateAuthenticationStatus response body cannot close: %+v", rspCloseErr)
+		}
+	}()
+
+	return nil
+}
+
+func (p *Processor) GenerateAuthDataProcedure(authInfoRequest models.AuthenticationInfoRequest, supiOrSuci string) (
+	response *models.AuthenticationInfoResult, problemDetails *models.ProblemDetails,
+) {
+	ctx, pd, err := udm_context.GetSelf().GetTokenCtx(models.ServiceName_NUDR_DR, models.NfType_UDR)
 	if err != nil {
-		problemDetail := "[Request Body] " + err.Error()
-		rsp := models.ProblemDetails{
-			Title:  "Malformed request syntax",
-			Status: http.StatusBadRequest,
-			Detail: problemDetail,
+		return nil, pd
+	}
+	logger.UeauLog.Traceln("In GenerateAuthDataProcedure")
+
+	response = &models.AuthenticationInfoResult{}
+	rand.New(rand.NewSource(time.Now().UnixNano()))
+	supi, err := suci.ToSupi(supiOrSuci, udm_context.GetSelf().SuciProfiles)
+	if err != nil {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  authenticationRejected,
+			Detail: err.Error(),
 		}
-		logger.UeauLog.Errorln(problemDetail)
-		c.JSON(http.StatusBadRequest, rsp)
-		return
+
+		logger.UeauLog.Errorln("suciToSupi error: ", err.Error())
+		return nil, problemDetails
 	}
 
-	// step 1: log
-	logger.UeauLog.Infoln("Handle GenerateAuthDataRequest")
+	logger.UeauLog.Tracef("supi conversion => [%s]", supi)
 
-	// step 2: retrieve request
-	supiOrSuci := c.Param("supiOrSuci")
+	client, err := p.consumer.CreateUDMClientToUDR(supi)
+	if err != nil {
+		return nil, openapi.ProblemDetailsSystemFailure(err.Error())
+	}
+	authSubs, res, err := client.AuthenticationDataDocumentApi.QueryAuthSubsData(ctx, supi, nil)
+	if err != nil {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  authenticationRejected,
+			Detail: err.Error(),
+		}
 
-	// step 3: handle the message
-	response, problemDetails := p.consumer.GenerateAuthDataProcedure(authInfoReq, supiOrSuci)
+		switch res.StatusCode {
+		case http.StatusNotFound:
+			logger.UeauLog.Warnf("Return from UDR QueryAuthSubsData error")
+		default:
+			logger.UeauLog.Errorln("Return from UDR QueryAuthSubsData error")
+		}
+		return nil, problemDetails
+	}
+	defer func() {
+		if rspCloseErr := res.Body.Close(); rspCloseErr != nil {
+			logger.SdmLog.Errorf("QueryAuthSubsData response body cannot close: %+v", rspCloseErr)
+		}
+	}()
 
-	// step 4: process the return value from step 3
-	if response != nil {
-		// status code is based on SPEC, and option headers
-		c.JSON(http.StatusOK, response)
-		return
-	} else if problemDetails != nil {
-		c.JSON(int(problemDetails.Status), problemDetails)
-		return
+	/*
+		K, RAND, CK, IK: 128 bits (16 bytes) (hex len = 32)
+		SQN, AK: 48 bits (6 bytes) (hex len = 12) TS33.102 - 6.3.2
+		AMF: 16 bits (2 bytes) (hex len = 4) TS33.102 - Annex H
+	*/
+
+	hasK, hasOP, hasOPC := false, false, false
+	var kStr, opStr, opcStr string
+	var k, op, opc []byte
+
+	if authSubs.PermanentKey != nil {
+		kStr = authSubs.PermanentKey.PermanentKeyValue
+		if len(kStr) == keyStrLen {
+			k, err = hex.DecodeString(kStr)
+			if err != nil {
+				logger.UeauLog.Errorln("err:", err)
+			} else {
+				hasK = true
+			}
+		} else {
+			problemDetails = &models.ProblemDetails{
+				Status: http.StatusForbidden,
+				Cause:  authenticationRejected,
+			}
+
+			logger.UeauLog.Errorln("kStr length is ", len(kStr))
+			return nil, problemDetails
+		}
 	} else {
 		problemDetails = &models.ProblemDetails{
 			Status: http.StatusForbidden,
-			Cause:  "UNSPECIFIED",
+			Cause:  authenticationRejected,
 		}
-		c.JSON(http.StatusForbidden, problemDetails)
-		return
+
+		logger.UeauLog.Errorln("Nil PermanentKey")
+		return nil, problemDetails
 	}
+
+	if authSubs.Milenage != nil {
+		if authSubs.Milenage.Op != nil && authSubs.Milenage.Op.OpValue != "" {
+			opStr = authSubs.Milenage.Op.OpValue
+			if len(opStr) == opStrLen {
+				op, err = hex.DecodeString(opStr)
+				if err != nil {
+					logger.UeauLog.Errorln("err:", err)
+				} else {
+					hasOP = true
+				}
+			} else {
+				logger.UeauLog.Errorln("opStr length is ", len(opStr))
+			}
+		} else {
+			logger.UeauLog.Infoln("Nil Op")
+		}
+	} else {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  authenticationRejected,
+		}
+
+		logger.UeauLog.Infoln("Nil Milenage")
+		return nil, problemDetails
+	}
+
+	if authSubs.Opc != nil && authSubs.Opc.OpcValue != "" {
+		opcStr = authSubs.Opc.OpcValue
+		if len(opcStr) == opcStrLen {
+			opc, err = hex.DecodeString(opcStr)
+			if err != nil {
+				logger.UeauLog.Errorln("err:", err)
+			} else {
+				hasOPC = true
+			}
+		} else {
+			logger.UeauLog.Errorln("opcStr length is ", len(opcStr))
+		}
+	} else {
+		logger.UeauLog.Infoln("Nil Opc")
+	}
+
+	if !hasOPC && !hasOP {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  authenticationRejected,
+		}
+
+		return nil, problemDetails
+	}
+
+	if !hasOPC {
+		if hasK && hasOP {
+			opc, err = milenage.GenerateOPC(k, op)
+			if err != nil {
+				logger.UeauLog.Errorln("milenage GenerateOPC err:", err)
+			}
+		} else {
+			problemDetails = &models.ProblemDetails{
+				Status: http.StatusForbidden,
+				Cause:  authenticationRejected,
+			}
+
+			logger.UeauLog.Errorln("Unable to derive OPC")
+			return nil, problemDetails
+		}
+	}
+
+	sqnStr := p.strictHex(authSubs.SequenceNumber, 12)
+	logger.UeauLog.Traceln("sqnStr", sqnStr)
+	sqn, err := hex.DecodeString(sqnStr)
+	if err != nil {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  authenticationRejected,
+			Detail: err.Error(),
+		}
+
+		logger.UeauLog.Errorln("err:", err)
+		return nil, problemDetails
+	}
+
+	logger.UeauLog.Tracef("K=[%x], sqn=[%x], OP=[%x], OPC=[%x]", k, sqn, op, opc)
+
+	RAND := make([]byte, 16)
+	_, err = cryptoRand.Read(RAND)
+	if err != nil {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  authenticationRejected,
+			Detail: err.Error(),
+		}
+
+		logger.UeauLog.Errorln("err:", err)
+		return nil, problemDetails
+	}
+
+	amfStr := p.strictHex(authSubs.AuthenticationManagementField, 4)
+	logger.UeauLog.Traceln("amfStr", amfStr)
+	AMF, err := hex.DecodeString(amfStr)
+	if err != nil {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  authenticationRejected,
+			Detail: err.Error(),
+		}
+
+		logger.UeauLog.Errorln("err:", err)
+		return nil, problemDetails
+	}
+
+	logger.UeauLog.Tracef("RAND=[%x], AMF=[%x]", RAND, AMF)
+
+	// re-synchronization
+	if authInfoRequest.ResynchronizationInfo != nil {
+		logger.UeauLog.Infof("Authentication re-synchronization")
+
+		Auts, deCodeErr := hex.DecodeString(authInfoRequest.ResynchronizationInfo.Auts)
+		if deCodeErr != nil {
+			problemDetails = &models.ProblemDetails{
+				Status: http.StatusForbidden,
+				Cause:  authenticationRejected,
+				Detail: deCodeErr.Error(),
+			}
+
+			logger.UeauLog.Errorln("err:", deCodeErr)
+			return nil, problemDetails
+		}
+
+		randHex, deCodeErr := hex.DecodeString(authInfoRequest.ResynchronizationInfo.Rand)
+		if deCodeErr != nil {
+			problemDetails = &models.ProblemDetails{
+				Status: http.StatusForbidden,
+				Cause:  authenticationRejected,
+				Detail: deCodeErr.Error(),
+			}
+
+			logger.UeauLog.Errorln("err:", deCodeErr)
+			return nil, problemDetails
+		}
+
+		SQNms, macS := p.aucSQN(opc, k, Auts, randHex)
+		if reflect.DeepEqual(macS, Auts[6:]) {
+			_, err = cryptoRand.Read(RAND)
+			if err != nil {
+				problemDetails = &models.ProblemDetails{
+					Status: http.StatusForbidden,
+					Cause:  authenticationRejected,
+					Detail: deCodeErr.Error(),
+				}
+
+				logger.UeauLog.Errorln("err:", deCodeErr)
+				return nil, problemDetails
+			}
+
+			// increment sqn authSubs.SequenceNumber
+			bigSQN := big.NewInt(0)
+			sqnStr = hex.EncodeToString(SQNms)
+			logger.UeauLog.Tracef("SQNstr=[%s]", sqnStr)
+			bigSQN.SetString(sqnStr, 16)
+
+			bigInc := big.NewInt(ind + 1)
+
+			bigP := big.NewInt(SqnMAx)
+			bigSQN = bigInc.Add(bigSQN, bigInc)
+			bigSQN = bigSQN.Mod(bigSQN, bigP)
+			sqnStr = fmt.Sprintf("%x", bigSQN)
+			sqnStr = p.strictHex(sqnStr, 12)
+		} else {
+			logger.UeauLog.Errorln("Re-Sync MAC failed ", supiOrSuci)
+			// Check if suci
+			suciPart := strings.Split(supiOrSuci, "-")
+			if suciPart[suci.PrefixPlace] == suci.PrefixSUCI &&
+				suciPart[suci.SupiTypePlace] == suci.SupiTypeIMSI &&
+				suciPart[suci.SchemePlace] != suci.NullScheme {
+				// Get SuciProfile index and write public key
+				keyIndex, err1 := strconv.Atoi(suciPart[suci.HNPublicKeyIDPlace])
+				if err1 != nil {
+					logger.UeauLog.Errorln("Re-Sync Failed UDM Public Key HNPublicKeyIDPlace parse Error")
+				} else if keyIndex < 1 {
+					logger.UeauLog.Errorf("Re-Sync Failed UDM Public Key HNPublicKeyIDPlace keyIndex[%d] < 1",
+						keyIndex)
+				} else {
+					logger.UeauLog.Errorln("Re-Sync Failed UDM Public Key ",
+						udm_context.GetSelf().SuciProfiles[keyIndex-1].PublicKey)
+				}
+			}
+			logger.UeauLog.Errorln("MACS ", macS)
+			logger.UeauLog.Errorln("Auts[6:] ", Auts[6:])
+			logger.UeauLog.Errorln("Sqn ", SQNms)
+			problemDetails = &models.ProblemDetails{
+				Status: http.StatusForbidden,
+				Cause:  "modification is rejected",
+			}
+			return nil, problemDetails
+		}
+	}
+
+	// increment sqn
+	bigSQN := big.NewInt(0)
+	sqn, err = hex.DecodeString(sqnStr)
+	if err != nil {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  authenticationRejected,
+			Detail: err.Error(),
+		}
+
+		logger.UeauLog.Errorln("err:", err)
+		return nil, problemDetails
+	}
+
+	bigSQN.SetString(sqnStr, 16)
+
+	bigInc := big.NewInt(1)
+	bigSQN = bigInc.Add(bigSQN, bigInc)
+
+	SQNheStr := fmt.Sprintf("%x", bigSQN)
+	SQNheStr = p.strictHex(SQNheStr, 12)
+	patchItemArray := []models.PatchItem{
+		{
+			Op:    models.PatchOperation_REPLACE,
+			Path:  "/sequenceNumber",
+			Value: SQNheStr,
+		},
+	}
+
+	var rsp *http.Response
+
+	rsp, err = client.AuthenticationDataDocumentApi.ModifyAuthentication(
+		ctx, supi, patchItemArray)
+	if err != nil {
+		problemDetails = &models.ProblemDetails{
+			Status: http.StatusForbidden,
+			Cause:  "modification is rejected ",
+			Detail: err.Error(),
+		}
+
+		logger.UeauLog.Errorln("update sqn error:", err)
+		return nil, problemDetails
+	}
+	defer func() {
+		if rspCloseErr := rsp.Body.Close(); rspCloseErr != nil {
+			logger.SdmLog.Errorf("ModifyAuthentication response body cannot close: %+v", rspCloseErr)
+		}
+	}()
+
+	// Run milenage
+	macA, macS := make([]byte, 8), make([]byte, 8)
+	CK, IK := make([]byte, 16), make([]byte, 16)
+	RES := make([]byte, 8)
+	AK, AKstar := make([]byte, 6), make([]byte, 6)
+
+	// Generate macA, macS
+	err = milenage.F1(opc, k, RAND, sqn, AMF, macA, macS)
+	if err != nil {
+		logger.UeauLog.Errorln("milenage F1 err:", err)
+	}
+
+	// Generate RES, CK, IK, AK, AKstar
+	// RES == XRES (expected RES) for server
+	err = milenage.F2345(opc, k, RAND, RES, CK, IK, AK, AKstar)
+	if err != nil {
+		logger.UeauLog.Errorln("milenage F2345 err:", err)
+	}
+	logger.UeauLog.Tracef("milenage RES=[%s]", hex.EncodeToString(RES))
+
+	// Generate AUTN
+	logger.UeauLog.Tracef("SQN=[%x], AK=[%x]", sqn, AK)
+	logger.UeauLog.Tracef("AMF=[%x], macA=[%x]", AMF, macA)
+	SQNxorAK := make([]byte, 6)
+	for i := 0; i < len(sqn); i++ {
+		SQNxorAK[i] = sqn[i] ^ AK[i]
+	}
+	logger.UeauLog.Tracef("SQN xor AK=[%x]", SQNxorAK)
+	AUTN := append(append(SQNxorAK, AMF...), macA...)
+	logger.UeauLog.Tracef("AUTN=[%x]", AUTN)
+
+	var av models.AuthenticationVector
+	if authSubs.AuthenticationMethod == models.AuthMethod__5_G_AKA {
+		response.AuthType = models.AuthType__5_G_AKA
+
+		// derive XRES*
+		key := append(CK, IK...)
+		FC := ueauth.FC_FOR_RES_STAR_XRES_STAR_DERIVATION
+		P0 := []byte(authInfoRequest.ServingNetworkName)
+		P1 := RAND
+		P2 := RES
+
+		kdfValForXresStar, err := ueauth.GetKDFValue(
+			key, FC, P0, ueauth.KDFLen(P0), P1, ueauth.KDFLen(P1), P2, ueauth.KDFLen(P2))
+		if err != nil {
+			logger.UeauLog.Errorf("Get kdfValForXresStar err: %+v", err)
+		}
+		xresStar := kdfValForXresStar[len(kdfValForXresStar)/2:]
+		logger.UeauLog.Tracef("xresStar=[%x]", xresStar)
+
+		// derive Kausf
+		FC = ueauth.FC_FOR_KAUSF_DERIVATION
+		P0 = []byte(authInfoRequest.ServingNetworkName)
+		P1 = SQNxorAK
+		kdfValForKausf, err := ueauth.GetKDFValue(key, FC, P0, ueauth.KDFLen(P0), P1, ueauth.KDFLen(P1))
+		if err != nil {
+			logger.UeauLog.Errorf("Get kdfValForKausf err: %+v", err)
+		}
+		logger.UeauLog.Tracef("Kausf=[%x]", kdfValForKausf)
+
+		// Fill in rand, xresStar, autn, kausf
+		av.Rand = hex.EncodeToString(RAND)
+		av.XresStar = hex.EncodeToString(xresStar)
+		av.Autn = hex.EncodeToString(AUTN)
+		av.Kausf = hex.EncodeToString(kdfValForKausf)
+		av.AvType = models.AvType__5_G_HE_AKA
+	} else { // EAP-AKA'
+		response.AuthType = models.AuthType_EAP_AKA_PRIME
+
+		// derive CK' and IK'
+		key := append(CK, IK...)
+		FC := ueauth.FC_FOR_CK_PRIME_IK_PRIME_DERIVATION
+		P0 := []byte(authInfoRequest.ServingNetworkName)
+		P1 := SQNxorAK
+		kdfVal, err := ueauth.GetKDFValue(key, FC, P0, ueauth.KDFLen(P0), P1, ueauth.KDFLen(P1))
+		if err != nil {
+			logger.UeauLog.Errorf("Get kdfVal err: %+v", err)
+		}
+		logger.UeauLog.Tracef("kdfVal=[%x] (len=%d)", kdfVal, len(kdfVal))
+
+		// For TS 35.208 test set 19 & RFC 5448 test vector 1
+		// CK': 0093 962d 0dd8 4aa5 684b 045c 9edf fa04
+		// IK': ccfc 230c a74f cc96 c0a5 d611 64f5 a76
+
+		ckPrime := kdfVal[:len(kdfVal)/2]
+		ikPrime := kdfVal[len(kdfVal)/2:]
+		logger.UeauLog.Tracef("ckPrime=[%x], kPrime=[%x]", ckPrime, ikPrime)
+
+		// Fill in rand, xres, autn, ckPrime, ikPrime
+		av.Rand = hex.EncodeToString(RAND)
+		av.Xres = hex.EncodeToString(RES)
+		av.Autn = hex.EncodeToString(AUTN)
+		av.CkPrime = hex.EncodeToString(ckPrime)
+		av.IkPrime = hex.EncodeToString(ikPrime)
+		av.AvType = models.AvType_EAP_AKA_PRIME
+	}
+
+	response.AuthenticationVector = &av
+	response.Supi = supi
+	return response, nil
 }
