--- internal/sbi/processor/index.go
@@ -1,11 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-func (p *Processor) HandleIndex(c *gin.Context) {
-	c.String(http.StatusOK, "Hello World!")
-}
