+++ internal/sbi/processor/parameter_provision.go
@@ -1,0 +1,37 @@
+package processor
+
+import (
+	"github.com/free5gc/openapi"
+	"github.com/free5gc/openapi/models"
+	udm_context "github.com/free5gc/udm/internal/context"
+	"github.com/free5gc/udm/internal/logger"
+)
+
+func (p *Processor) UpdateProcedure(
+	updateRequest models.PpData,
+	gpsi string,
+) (problemDetails *models.ProblemDetails) {
+	ctx, pd, err := udm_context.GetSelf().GetTokenCtx(models.ServiceName_NUDR_DR, models.NfType_UDR)
+	if err != nil {
+		return pd
+	}
+	clientAPI, err := p.consumer.CreateUDMClientToUDR(gpsi)
+	if err != nil {
+		return openapi.ProblemDetailsSystemFailure(err.Error())
+	}
+	res, err := clientAPI.ProvisionedParameterDataDocumentApi.ModifyPpData(ctx, gpsi, nil)
+	if err != nil {
+		problemDetails = &models.ProblemDetails{
+			Status: int32(res.StatusCode),
+			Cause:  err.(openapi.GenericOpenAPIError).Model().(models.ProblemDetails).Cause,
+			Detail: err.Error(),
+		}
+		return problemDetails
+	}
+	defer func() {
+		if rspCloseErr := res.Body.Close(); rspCloseErr != nil {
+			logger.PpLog.Errorf("ModifyPpData response body cannot close: %+v", rspCloseErr)
+		}
+	}()
+	return nil
+}
