--- internal/sbi/processor/parameter_update_in_the_amf_registration_for3_gpp_access.go
@@ -1,62 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-
-	"github.com/free5gc/openapi"
-	"github.com/free5gc/openapi/models"
-	"github.com/free5gc/udm/internal/logger"
-)
-
-// UpdateAmf3gppAccess - Update a parameter in the AMF registration for 3GPP access
-func (p *Processor) HandleUpdateAmf3gppAccess(c *gin.Context) {
-	var amf3GppAccessRegistrationModification models.Amf3GppAccessRegistrationModification
-
-	// step 1: retrieve http request body
-	requestBody, err := c.GetRawData()
-	if err != nil {
-		problemDetail := models.ProblemDetails{
-			Title:  "System failure",
-			Status: http.StatusInternalServerError,
-			Detail: err.Error(),
-			Cause:  "SYSTEM_FAILURE",
-		}
-		logger.UecmLog.Errorf("Get Request Body error: %+v", err)
-		c.JSON(http.StatusInternalServerError, problemDetail)
-		return
-	}
-
-	// step 2: convert requestBody to openapi models
-	err = openapi.Deserialize(&amf3GppAccessRegistrationModification, requestBody, "application/json")
-	if err != nil {
-		problemDetail := "[Request Body] " + err.Error()
-		rsp := models.ProblemDetails{
-			Title:  "Malformed request syntax",
-			Status: http.StatusBadRequest,
-			Detail: problemDetail,
-		}
-		logger.UecmLog.Errorln(problemDetail)
-		c.JSON(http.StatusBadRequest, rsp)
-		return
-	}
-
-	// step 1: log
-	logger.UecmLog.Infof("Handle UpdateAmf3gppAccessRequest")
-
-	// step 2: retrieve request
-	ueID := c.Param("ueId")
-
-	// step 3: handle the message
-	problemDetails := p.consumer.UpdateAmf3gppAccessProcedure(amf3GppAccessRegistrationModification, ueID)
-
-	// step 4: process the return value from step 3
-	if problemDetails != nil {
-		c.JSON(int(problemDetails.Status), problemDetails)
-		return
-	} else {
-		c.Status(http.StatusNoContent)
-		return
-	}
-}
