--- internal/sbi/processor/providing_acknowledgement_of_steering_of_roaming.go
@@ -1,12 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-// Info - Nudm_Sdm Info service operation
-func (p *Processor) HandleInfo(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{})
-}
