--- internal/sbi/processor/providing_acknowledgement_of_ue_parameters_update.go
@@ -1,12 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-// PutUpuAck - Nudm_Sdm Info for UPU service operation
-func (p *Processor) HandlePutUpuAck(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{})
-}
