--- internal/sbi/processor/retrieval_of_multiple_data_sets.go
@@ -1,54 +1,0 @@
-package processor
-
-import (
-	"net/http"
-	"net/url"
-	"strings"
-
-	"github.com/gin-gonic/gin"
-
-	"github.com/free5gc/openapi/models"
-	"github.com/free5gc/udm/internal/logger"
-)
-
-// GetSupi - retrieve multiple data sets
-func (p *Processor) HandleGetSupi(c *gin.Context) {
-	var query url.Values
-	query.Set("plmn-id", c.Query("plmn-id"))
-	query.Set("dataset-names", c.Query("dataset-names"))
-	query.Set("supported-features", c.Query("supported-features"))
-
-	// step 1: log
-	logger.SdmLog.Infof("Handle GetSupiRequest")
-
-	// step 2: retrieve request
-	supi := c.Params.ByName("supi")
-	plmnIDStruct, problemDetails := p.getPlmnIDStruct(query)
-	if problemDetails != nil {
-		c.JSON(int(problemDetails.Status), problemDetails)
-		return
-	}
-	plmnID := plmnIDStruct.Mcc + plmnIDStruct.Mnc
-	dataSetNames := strings.Split(query.Get("dataset-names"), ",")
-	supportedFeatures := query.Get("supported-features")
-
-	// step 3: handle the message
-	response, problemDetails := p.consumer.GetSupiProcedure(supi, plmnID, dataSetNames, supportedFeatures)
-
-	// step 4: process the return value from step 3
-	if response != nil {
-		// status code is based on SPEC, and option headers
-		c.JSON(http.StatusOK, response)
-		return
-	} else if problemDetails != nil {
-		c.JSON(int(problemDetails.Status), problemDetails)
-		return
-	} else {
-		problemDetails = &models.ProblemDetails{
-			Status: http.StatusForbidden,
-			Cause:  "UNSPECIFIED",
-		}
-		c.JSON(http.StatusForbidden, problemDetails)
-		return
-	}
-}
