--- internal/sbi/processor/sms_management_subscription_data_retrieval.go
@@ -1,12 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-// GetSmsMngData - retrieve a UE's SMS Management Subscription Data
-func (p *Processor) HandleGetSmsMngData(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{})
-}
