--- internal/sbi/processor/sms_subscription_data_retrieval.go
@@ -1,12 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-// GetSmsData - retrieve a UE's SMS Subscription Data
-func (p *Processor) HandleGetSmsData(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{})
-}
