--- internal/sbi/processor/smsf_deregistration_for_non3_gpp_access.go
@@ -1,12 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-// DeregistrationSmsfNon3gppAccess - delete SMSF registration for non 3GPP access
-func (p *Processor) HandleDeregistrationSmsfNon3gppAccess(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{})
-}
