--- internal/sbi/processor/smsf_non3_gpp_access_registration_info_retrieval.go
@@ -1,12 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-// GetSmsfNon3gppAccess - retrieve the SMSF registration for non-3GPP access information
-func (p *Processor) HandleGetSmsfNon3gppAccess(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{})
-}
