--- internal/sbi/processor/smsf_registration_for3_gpp_access.go
@@ -1,12 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-// UpdateSMSFReg3GPP - register as SMSF for 3GPP access
-func (p *Processor) HandleUpdateSMSFReg3GPP(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{})
-}
