--- internal/sbi/processor/subscription_deletion_for_shared_data.go
@@ -1,28 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-
-	"github.com/free5gc/udm/internal/logger"
-)
-
-// UnsubscribeForSharedData - unsubscribe from notifications for shared data
-func (p *Processor) HandleUnsubscribeForSharedData(c *gin.Context) {
-	logger.SdmLog.Infof("Handle UnsubscribeForSharedData")
-
-	// step 2: retrieve request
-	subscriptionID := c.Params.ByName("subscriptionId")
-	// step 3: handle the message
-	problemDetails := p.consumer.UnsubscribeForSharedDataProcedure(subscriptionID)
-
-	// step 4: process the return value from step 3
-	if problemDetails != nil {
-		c.JSON(int(problemDetails.Status), problemDetails)
-		return
-	} else {
-		c.Status(http.StatusNoContent)
-		return
-	}
-}
