--- internal/sbi/processor/subscription_modification.go
@@ -1,129 +1,0 @@
-package processor
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-
-	"github.com/free5gc/openapi"
-	"github.com/free5gc/openapi/models"
-	"github.com/free5gc/udm/internal/logger"
-)
-
-// Modify - modify the subscription
-func (p *Processor) HandleModify(c *gin.Context) {
-	var sdmSubsModificationReq models.SdmSubsModification
-	// step 1: retrieve http request body
-	requestBody, err := c.GetRawData()
-	if err != nil {
-		problemDetail := models.ProblemDetails{
-			Title:  "System failure",
-			Status: http.StatusInternalServerError,
-			Detail: err.Error(),
-			Cause:  "SYSTEM_FAILURE",
-		}
-		logger.SdmLog.Errorf("Get Request Body error: %+v", err)
-		c.JSON(http.StatusInternalServerError, problemDetail)
-		return
-	}
-
-	// step 2: convert requestBody to openapi models
-	err = openapi.Deserialize(&sdmSubsModificationReq, requestBody, "application/json")
-	if err != nil {
-		problemDetail := "[Request Body] " + err.Error()
-		rsp := models.ProblemDetails{
-			Title:  "Malformed request syntax",
-			Status: http.StatusBadRequest,
-			Detail: problemDetail,
-		}
-		logger.SdmLog.Errorln(problemDetail)
-		c.JSON(http.StatusBadRequest, rsp)
-		return
-	}
-
-	// step 1: log
-	logger.SdmLog.Infof("Handle Modify")
-
-	// step 2: retrieve request
-	supi := c.Params.ByName("supi")
-	subscriptionID := c.Params.ByName("subscriptionId")
-
-	// step 3: handle the message
-	response, problemDetails := p.consumer.ModifyProcedure(&sdmSubsModificationReq, supi, subscriptionID)
-
-	// step 4: process the return value from step 3
-	if response != nil {
-		// status code is based on SPEC, and option headers
-		c.JSON(http.StatusOK, response)
-		return
-	} else if problemDetails != nil {
-		c.JSON(int(problemDetails.Status), problemDetails)
-		return
-	} else {
-		problemDetails = &models.ProblemDetails{
-			Status: http.StatusForbidden,
-			Cause:  "UNSPECIFIED",
-		}
-		c.JSON(http.StatusForbidden, problemDetails)
-		return
-	}
-}
-
-// ModifyForSharedData - modify the subscription
-func (p *Processor) HandleModifyForSharedData(c *gin.Context) {
-	var sharedDataSubscriptions models.SdmSubsModification
-	// step 1: retrieve http request body
-	requestBody, err := c.GetRawData()
-	if err != nil {
-		problemDetail := models.ProblemDetails{
-			Title:  "System failure",
-			Status: http.StatusInternalServerError,
-			Detail: err.Error(),
-			Cause:  "SYSTEM_FAILURE",
-		}
-		logger.SdmLog.Errorf("Get Request Body error: %+v", err)
-		c.JSON(http.StatusInternalServerError, problemDetail)
-		return
-	}
-
-	// step 2: convert requestBody to openapi models
-	err = openapi.Deserialize(&sharedDataSubscriptions, requestBody, "application/json")
-	if err != nil {
-		problemDetail := "[Request Body] " + err.Error()
-		rsp := models.ProblemDetails{
-			Title:  "Malformed request syntax",
-			Status: http.StatusBadRequest,
-			Detail: problemDetail,
-		}
-		logger.SdmLog.Errorln(problemDetail)
-		c.JSON(http.StatusBadRequest, rsp)
-		return
-	}
-
-	// step 1: log
-	logger.SdmLog.Infof("Handle ModifyForSharedData")
-
-	// step 2: retrieve request
-	supi := c.Params.ByName("supi")
-	subscriptionID := c.Params.ByName("subscriptionId")
-
-	// step 3: handle the message
-	response, problemDetails := p.consumer.ModifyForSharedDataProcedure(&sharedDataSubscriptions, supi, subscriptionID)
-
-	// step 4: process the return value from step 3
-	if response != nil {
-		// status code is based on SPEC, and option headers
-		c.JSON(http.StatusOK, response)
-		return
-	} else if problemDetails != nil {
-		c.JSON(int(problemDetails.Status), problemDetails)
-		return
-	} else {
-		problemDetails = &models.ProblemDetails{
-			Status: http.StatusForbidden,
-			Cause:  "UNSPECIFIED",
-		}
-		c.JSON(http.StatusForbidden, problemDetails)
-		return
-	}
-}
